Topic and Subtopics controllers, routers and models are seperated
