+++ src/routes/subtopicRouter.ts
@@ -1,0 +1,9 @@
+import express from "express";
+const router = express.Router();
+
+import { getSubtopics, addSubtopic } from "../controllers/subtopicController";
+
+router.get("/", getSubtopics);
+router.post("/", addSubtopic);
+
+export default router;
